refactor : LottoTicekt의 중복 로직 제거로 인한 변경 및 보너스 번호를 int로 받던 로직을 LottoNumber 객체로 받도록 변경
